TEX-107. RUn tool window UI and executor WIP: compile steps are now integrated.
As a workaround, a composite process handler is constructed that executes all compile steps in sequence. Steps are shown in the build view, as well as (some) log messages.

Next steps: capture exit codes as well as the actual output, and handle iterruptions in the UI to actually stop the whole execution.